refactor: rename multi-select-autocomplete to combobox for consistency
