+++ libs/extensions/angular/combobox/src/combobox.component.scss
