+++ libs/extensions/angular/combobox/src/index.ts
@@ -1,0 +1,1 @@
+export { ComboboxComponent } from './combobox.component';
