--- libs/extensions/angular/multi-select-autocomplete/src/index.ts
@@ -1,1 +1,0 @@
-export { MultiSelectAutocompleteComponent } from './multi-select-autocomplete.component';
