+++ tsconfig.base.json
@@ -80,9 +80,7 @@
       "@kirbydesign/extensions-angular/spot-illustration": [
         "libs/extensions/angular/dist/spot-illustration"
       ],
-      "@kirbydesign/extensions-angular/multi-select-autocomplete": [
-        "libs/extensions/angular/dist/multi-select-autocomplete"
-      ],
+      "@kirbydesign/extensions-angular/combobox": ["libs/extensions/angular/dist/combobox"],
       "@repo/docs-import-map/*": ["tools/docs-import-map/*"],
       "~/app/*": ["apps/cookbook/src/app/*"],
       "~/environments/*": ["apps/cookbook/src/environments/*"]
